Add mixin support
Mixins are functions or objects which can be applied to any component to
modify or override its behavior.  Mixable components are componets which can
either be used alone, or used as mixin with other components.  elements. This
allows better representation of the MDL component model, where almost any MDL
component or even plain HTML element can be a grid container or a grid cell
and <a> tags can be buttons, for instance.